Remove docset dropdown menu from all pages
