+++ docs/plugins/gatsby-theme-flow/gatsby-node.js
@@ -254,6 +254,7 @@ async function createPagesForSection(
         id,
         subtitle,
         sidebar: {
+          showMainNav: section.sidebarShowMainNav,
           alwaysExpanded: section.sidebarAlwaysExpanded,
           contents: sidebarContents
         },
