+++ docs/plugins/gatsby-theme-flow/index.js
@@ -1,11 +1,13 @@
-export {MultiCodeBlock} from './src/components/multi-code-block';
+export { MultiCodeBlock } from './src/components/multi-code-block';
+
 export {
   ExpansionPanel,
   ExpansionPanelList,
   ExpansionPanelListItem
 } from './src/components/expansion-panel';
-export {NavItemsContext} from './src/components/page-layout';
+
 export {
+  NavItemTitle,
   NavItemDescription,
-  NavItemTitle
-} from './src/components/docset-switcher';
+  NavItemsContext,
+} from './src/components/page-layout';
