+++ docs/plugins/gatsby-theme-flow/src/utils/typography.js
@@ -1,4 +1,4 @@
 export const smallCaps = {
-  letterSpacing: '0.142em',
+  letterSpacing: '0.1em',
   textTransform: 'uppercase'
 };
